+++ packages/connect-wallet-modal/src/components/WalletsModal/WalletsModalForEth.tsx
@@ -5,6 +5,7 @@ import {
   ConnectWC,
   ConnectCoinbase,
   ConnectBrowser,
+  ConnectBinance,
 } from '../../connectButtons';
 import { WalletsModal } from './WalletsModal';
 import type { WalletIdsEthereum } from '@reef-knot/wallets-list';
@@ -19,17 +20,19 @@ const buttonComponentsByConnectorId: WalletsModalEthProps['buttonComponentsByCon
     walletConnect: ConnectWC,
     coinbaseWallet: ConnectCoinbase,
     ledgerHID: ConnectLedger,
+    binanceWallet: ConnectBinance,
   };
 
 const WALLETS_DISPLAY_CONFIG_DEFAULT: WalletsModalEthProps['walletsShown'] = [
   'browserExtension',
   'metaMask',
+  'okx',
   'ledgerHID',
   'ledgerLive',
   'walletConnect',
+  'binanceWallet',
   'coinbase',
   'trust',
-  'okx',
   'exodus',
   'brave',
   'bitget',